# docs/swagger.yaml
definitions:
  models.Book:
    properties:
      author:
        type: string
      id:
        type: integer
      price:
        type: number
      purchased:
        type: integer
      title:
        type: string
    type: object
  pkg.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  pkg.MessageResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления книгами.
  title: Book API
  version: "1.0"
paths:
  /books:
    get:
      consumes:
      - application/json
      description: Возвращает массив книг с возможностью пагинации.
      parameters:
      - default: 0
        description: Смещение
        in: query
        name: skip
        type: integer
      - default: 10
        description: Лимит
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Book'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Получить список книг
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Добавляет новую книгу в базу данных.
      parameters:
      - description: Данные книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Создать новую книгу
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет книгу из базы данных по её ID.
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Удалить книгу
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Возвращает информацию о книге по её ID.
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Получить информацию о книге
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Обновляет информацию о книге в базе данных.
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      - description: Данные книги
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Book'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ErrorResponse'
      summary: Обновить информацию о книге
      tags:
      - books
swagger: "2.0"